requirements.txt
alembic==1.4.1
analytics-python==1.3.1
appdirs==1.4.4
astor==0.8.1
atpublic==2.3
Babel==2.9.1
backoff==1.10.0
bce-python-sdk==0.8.61
bcrypt==3.2.0
cached-property==1.5.2
certifi==2021.5.30
cffi==1.14.5
cfgv==3.3.0
chardet==4.0.0
click==8.0.1
cloudpickle==1.6.0
colorama==0.4.4
colorlog==5.0.1
commonmark==0.9.1
configobj==5.0.6
cryptography==3.4.7
cycler==0.10.0
dask==2021.3.0
databricks-cli==0.14.3
dataclasses==0.7
decorator==4.4.2
dictdiffer==0.8.1
dill==0.3.4
diskcache==5.2.1
distlib==0.3.2
distro==1.5.0
docker==5.0.0
dpath==2.0.1
dulwich==0.20.23
dvc==2.5.0
easydict==1.9
entrypoints==0.3
ffmpy==0.3.0
filelock==3.0.12
flake8==3.9.2
Flask==2.0.1
Flask-Babel==2.0.0
Flask-CacheBuster==1.0.0
Flask-Cors==3.0.10
Flask-Login==0.5.0
flatten-dict==0.4.0
flufl.lock==3.2
fsspec==2021.6.1
ftfy==6.0.3
funcy==1.16
future==0.18.2
gast==0.5.0
gitdb==4.0.7
GitPython==3.1.18
gradio==2.1.2
grandalf==0.6
greenlet==1.1.0
gunicorn==20.1.0
h5py==3.1.0
identify==2.2.10
idna==2.10
importlib-metadata==4.6.1
importlib-resources==5.2.0
invoke==1.5.0
itsdangerous==2.0.1
jieba==0.42.1
Jinja2==3.0.1
joblib==1.0.1
jsonpath-ng==1.5.3
kiwisolver==1.3.1
mailchecker==4.0.10
Mako==1.1.4
markdown2==2.4.0
MarkupSafe==2.0.1
matplotlib==3.3.4
mccabe==0.6.1
mlflow==1.18.0
monotonic==1.6
multiprocess==0.70.12.2
nanotime==0.5.2
networkx==2.5.1
nodeenv==1.6.0
numpy==1.15.2
opencv-python==4.5.2.54
packaging==21.0
paddle2onnx==0.6
paddlehub==2.1.0
paddlenlp==2.0.5
paddlepaddle==2.1.1
pandas==1.0.5
paramiko==2.7.2
pathspec==0.8.1
phonenumbers==8.12.26
Pillow==8.1.1
pkg-resources==0.0.0
ply==3.11
pre-commit==2.13.0
prometheus-client==0.11.0
prometheus-flask-exporter==0.18.2
protobuf==3.17.3
psutil==5.8.0
pyasn1==0.4.8
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.10.1
pydot==1.4.2
pyflakes==2.3.1
pygit2==1.6.1
Pygments==2.9.0
pygtrie==2.4.2
PyNaCl==1.4.0
pyparsing==2.4.7
python-benedict==0.24.0
python-dateutil==2.8.1
python-editor==1.0.4
python-fsutil==0.5.0
python-slugify==5.0.2
pytz==2021.1
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==22.1.0
querystring-parser==1.2.4
rarfile==4.0
requests==2.25.1
rich==10.5.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
scikit-image==0.14.0
scikit-learn==0.24.2
scipy==1.5.4
seqeval==1.2.2
shellcheck-py==0.7.2.1
shortuuid==1.0.1
shtab==1.3.7
six==1.16.0
smmap==4.0.0
SQLAlchemy==1.4.20
sqlparse==0.4.1
tabulate==0.8.9
text-unidecode==1.3
threadpoolctl==2.1.0
toml==0.10.2
toolz==0.11.1
torch==1.9.0
torchvision==0.10.0
tqdm==4.61.2
typing-extensions==3.10.0.0
urllib3==1.26.6
virtualenv==20.4.7
visualdl==2.2.0
voluptuous==0.12.1
wcwidth==0.2.5
websocket-client==1.1.0
Werkzeug==2.0.1
xmltodict==0.12.0
zc.lockfile==2.0
zipp==3.5.0
